# docs/swagger/swagger.yaml
basePath: /
definitions:
  domains.Brand:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      website:
        type: string
    type: object
  domains.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  domains.Error:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing products, categories, brands, and skin types in an
    e-commerce system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-commerce API
  version: "1.0"
paths:
  /brands:
    get:
      consumes:
      - application/json
      description: Get a list of all brands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domains.Brand'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Get all brands
      tags:
      - brands
    post:
      consumes:
      - application/json
      description: Create a new brand with the provided details
      parameters:
      - description: Brand object
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/domains.Brand'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domains.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Create a new brand
      tags:
      - brands
  /brands/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a brand by its ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Delete brand
      tags:
      - brands
    get:
      consumes:
      - application/json
      description: Get a brand by its ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Get brand by ID
      tags:
      - brands
    put:
      consumes:
      - application/json
      description: Update an existing brand
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      - description: Brand object
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/domains.Brand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Update brand
      tags:
      - brands
  /categories:
    get:
      consumes:
      - application/json
      description: Get a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domains.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category with the provided details
      parameters:
      - description: Category object
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/domains.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domains.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Delete category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category object
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/domains.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Update category
      tags:
      - categories
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"